+++ gateway/gateway_test.go
@@ -2,7 +2,11 @@ package gateway
 
 import (
 	"context"
+	"encoding/json"
+	"fmt"
 	"gateway/node"
+	"os"
+	"path/filepath"
 	"testing"
 
 	"go.viam.com/rdk/components/sensor"
@@ -12,6 +16,53 @@ import (
 	"go.viam.com/test"
 	"go.viam.com/utils/protoutils"
 )
+
+func createDataFile(t *testing.T) *os.File {
+	// Create a temp device data file for testing
+	tmpDir := t.TempDir()
+	filePath := filepath.Join(tmpDir, "devices.txt")
+	file, err := os.OpenFile(filePath, os.O_RDWR|os.O_CREATE, 0o644)
+	test.That(t, err, test.ShouldBeNil)
+	return file
+}
+
+// setupTestGatewayWithFileDevice creates a test gateway with device info only in file.
+func setupFileAndGateway(t *testing.T) *gateway {
+	// Create a temp device data file for testing
+	tmpDir := t.TempDir()
+	filePath := filepath.Join(tmpDir, "devices.txt")
+	file, err := os.OpenFile(filePath, os.O_RDWR|os.O_CREATE, 0o644)
+	test.That(t, err, test.ShouldBeNil)
+
+	// Write device info to file
+	devices := []deviceInfo{
+		{
+			DevEUI:  fmt.Sprintf("%X", testDevEUI),
+			DevAddr: fmt.Sprintf("%X", testDeviceAddr),
+			AppSKey: fmt.Sprintf("%X", testAppSKey),
+		},
+	}
+	data, err := json.MarshalIndent(devices, "", "  ")
+	test.That(t, err, test.ShouldBeNil)
+	_, err = file.Write(data)
+	test.That(t, err, test.ShouldBeNil)
+
+	// Create gateway with empty devices map but device info in file
+	testDevices := make(map[string]*node.Node)
+	testNode := &node.Node{
+		NodeName:    testNodeName,
+		DecoderPath: testDecoderPath,
+		JoinType:    "OTAA",
+		DevEui:      testDevEUI,
+	}
+	testDevices[testNodeName] = testNode
+
+	return &gateway{
+		logger:   logging.NewTestLogger(t),
+		devices:  testDevices,
+		dataFile: file,
+	}
+}
 
 func TestValidate(t *testing.T) {
 	// Test valid config with default bus
@@ -63,9 +114,7 @@ func TestValidate(t *testing.T) {
 }
 
 func TestDoCommand(t *testing.T) {
-	g := &gateway{
-		devices: make(map[string]*node.Node),
-	}
+	g := setupFileAndGateway(t)
 
 	s := (sensor.Sensor)(g)
 
@@ -94,7 +143,7 @@ func TestDoCommand(t *testing.T) {
 		test.That(t, err, test.ShouldBeNil)
 	}
 
-	// Test Register device command
+	// Test Register device command - device not in file
 	registerCmd := make(map[string]interface{})
 	registerCmd["register_device"] = n
 	doOverWire(s, registerCmd)
@@ -115,6 +164,22 @@ func TestDoCommand(t *testing.T) {
 	test.That(t, dev.AppKey, test.ShouldResemble, testAppKey)
 	test.That(t, dev.DevEui, test.ShouldResemble, testDevEUI)
 	test.That(t, dev.DecoderPath, test.ShouldResemble, "/newpath")
+
+	// Test that if device is in file, OTAA fields get populated
+	dev = g.devices[testNodeName]
+	// clear map for the new test
+	g.devices = map[string]*node.Node{}
+	registerCmd["register_device"] = dev
+	doOverWire(s, registerCmd)
+	test.That(t, len(g.devices), test.ShouldEqual, 1)
+	dev, ok = g.devices[testNodeName]
+	test.That(t, ok, test.ShouldBeTrue)
+	test.That(t, dev.AppKey, test.ShouldResemble, testAppKey)
+	test.That(t, dev.DevEui, test.ShouldResemble, testDevEUI)
+	test.That(t, dev.DecoderPath, test.ShouldResemble, "/newpath")
+	// The dev addr and AppsKey should also be populated
+	test.That(t, dev.AppSKey, test.ShouldResemble, testAppSKey)
+	test.That(t, dev.Addr, test.ShouldResemble, testDeviceAddr)
 
 	// Test remove device command
 	removeCmd := make(map[string]interface{})
@@ -329,6 +394,70 @@ func TestStartCLogging(t *testing.T) {
 	test.That(t, loggingRoutineStarted["test-gateway"], test.ShouldBeTrue)
 }
 
+func TestSearchForDeviceInFile(t *testing.T) {
+	g := setupTestGateway(t)
+
+	// Device found in file should return device info
+	devices := []deviceInfo{
+		{
+			DevEUI:  fmt.Sprintf("%X", testDevEUI),
+			DevAddr: fmt.Sprintf("%X", testDeviceAddr),
+			AppSKey: "5572404C694E6B4C6F526132303138323",
+		},
+	}
+	data, err := json.MarshalIndent(devices, "", "  ")
+	test.That(t, err, test.ShouldBeNil)
+	_, err = g.dataFile.Write(data)
+	test.That(t, err, test.ShouldBeNil)
+
+	device, err := g.searchForDeviceInFile(testDevEUI)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, device, test.ShouldNotBeNil)
+	test.That(t, device.DevAddr, test.ShouldEqual, fmt.Sprintf("%X", testDeviceAddr))
+
+	//  Device not found in file should return errNoDevice
+	unknownAddr := []byte{0x01, 0x02, 0x03, 0x04}
+	device, err = g.searchForDeviceInFile(unknownAddr)
+	test.That(t, err, test.ShouldBeError, errNoDevice)
+	test.That(t, device, test.ShouldBeNil)
+
+	// Test File read error
+	g.dataFile.Close()
+	_, err = g.searchForDeviceInFile(testDeviceAddr)
+	test.That(t, err, test.ShouldNotBeNil)
+	test.That(t, err.Error(), test.ShouldContainSubstring, "failed to read device info from file")
+}
+
+func TestUpdateDeviceInfo(t *testing.T) {
+	g := setupTestGateway(t)
+
+	newAppSKey := []byte{
+		0x55, 0x72, 0x40, 0x4C,
+		0x69, 0x6E, 0x6B, 0x4C,
+		0x6F, 0x52, 0x61, 0x32,
+		0x31, 0x30, 0x32, 0x23,
+	}
+
+	newDevAddr := []byte{0xe2, 0x73, 0x65, 0x67}
+
+	// Test 1: Successful update
+	validInfo := &deviceInfo{
+		DevEUI:  fmt.Sprintf("%X", testDevEUI), // matches dev EUI on the gateway map
+		DevAddr: fmt.Sprintf("%X", newDevAddr),
+		AppSKey: fmt.Sprintf("%X", newAppSKey),
+	}
+
+	device := g.devices[testNodeName]
+
+	err := g.updateDeviceInfo(device, validInfo)
+	test.That(t, err, test.ShouldBeNil)
+	test.That(t, device, test.ShouldNotBeNil)
+	test.That(t, device.NodeName, test.ShouldEqual, testNodeName)
+	test.That(t, device.AppSKey, test.ShouldResemble, newAppSKey)
+	test.That(t, device.Addr, test.ShouldResemble, newDevAddr)
+
+}
+
 func TestClose(t *testing.T) {
 	// Create a gateway instance for testing
 	cfg := resource.Config{
@@ -370,6 +499,11 @@ func TestClose(t *testing.T) {
 	if g.logReader != nil {
 		buf := make([]byte, 1)
 		_, err = g.logReader.Read(buf)
+		test.That(t, err, test.ShouldNotBeNil)
+	}
+	if g.dataFile != nil {
+		buf := make([]byte, 1)
+		_, err = g.dataFile.Read(buf)
 		test.That(t, err, test.ShouldNotBeNil)
 	}
 }
