+++ gateway/uplink_test.go
@@ -2,13 +2,12 @@ package gateway
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
 	"os"
 	"path/filepath"
 	"testing"
 
 	"gateway/node"
+
 	"go.viam.com/rdk/logging"
 	"go.viam.com/test"
 )
@@ -60,44 +59,6 @@ func setupTestGateway(t *testing.T) *gateway {
 	testNode := &node.Node{
 		Addr:        testDeviceAddr,
 		AppSKey:     testAppSKey,
-		NodeName:    testNodeName,
-		DecoderPath: testDecoderPath,
-		JoinType:    "OTAA",
-		DevEui:      testDevEUI,
-	}
-	testDevices[testNodeName] = testNode
-
-	return &gateway{
-		logger:   logging.NewTestLogger(t),
-		devices:  testDevices,
-		dataFile: file,
-	}
-}
-
-// setupTestGatewayWithFileDevice creates a test gateway with device info only in file.
-func setupFileTest(t *testing.T) *gateway {
-	// Create a temp device data file for testing
-	tmpDir := t.TempDir()
-	filePath := filepath.Join(tmpDir, "devices.txt")
-	file, err := os.OpenFile(filePath, os.O_RDWR|os.O_CREATE, 0o644)
-	test.That(t, err, test.ShouldBeNil)
-
-	// Write device info to file
-	devices := []deviceInfo{
-		{
-			DevEUI:  fmt.Sprintf("%X", testDevEUI),
-			DevAddr: fmt.Sprintf("%X", testDeviceAddr),
-			AppSKey: fmt.Sprintf("%X", testAppSKey),
-		},
-	}
-	data, err := json.MarshalIndent(devices, "", "  ")
-	test.That(t, err, test.ShouldBeNil)
-	_, err = file.Write(data)
-	test.That(t, err, test.ShouldBeNil)
-
-	// Create gateway with empty devices map but device info in file
-	testDevices := make(map[string]*node.Node)
-	testNode := &node.Node{
 		NodeName:    testNodeName,
 		DecoderPath: testDecoderPath,
 		JoinType:    "OTAA",
@@ -182,89 +143,6 @@ func TestParseDataUplink(t *testing.T) {
 	test.That(t, err, test.ShouldNotBeNil)
 	test.That(t, err, test.ShouldBeError, errNoDevice)
 	g.Close(context.Background())
-
-	// Test that a device that is not in the device map but is in the persistent data file is added to the device map.
-	g = setupFileTest(t)
-	deviceName, readings, err = g.parseDataUplink(context.Background(), validUplinkData)
-	test.That(t, err, test.ShouldBeNil)
-	test.That(t, readings, test.ShouldNotBeNil)
-	test.That(t, deviceName, test.ShouldEqual, testNodeName)
-
-	// Verify device info was updated in device map
-	device, ok := g.devices[testNodeName]
-	test.That(t, ok, test.ShouldBeTrue)
-	test.That(t, device.Addr, test.ShouldResemble, testDeviceAddr)
-	test.That(t, len(device.AppSKey), test.ShouldEqual, 16)
-	test.That(t, device.AppSKey, test.ShouldResemble, testAppSKey)
-
-	g.Close(context.Background())
-}
-
-func TestSearchForDeviceInFile(t *testing.T) {
-	g := setupTestGateway(t)
-
-	// Device found in file should return device info
-	devices := []deviceInfo{
-		{
-			DevEUI:  "0102030405060708",
-			DevAddr: fmt.Sprintf("%X", testDeviceAddr),
-			AppSKey: "5572404C694E6B4C6F526132303138323",
-		},
-	}
-	data, err := json.MarshalIndent(devices, "", "  ")
-	test.That(t, err, test.ShouldBeNil)
-	_, err = g.dataFile.Write(data)
-	test.That(t, err, test.ShouldBeNil)
-
-	device, err := g.searchForDeviceInFile(testDeviceAddr)
-	test.That(t, err, test.ShouldBeNil)
-	test.That(t, device, test.ShouldNotBeNil)
-	test.That(t, device.DevAddr, test.ShouldEqual, fmt.Sprintf("%X", testDeviceAddr))
-
-	//  Device not found in file should return errNoDevice
-	unknownAddr := []byte{0x01, 0x02, 0x03, 0x04}
-	device, err = g.searchForDeviceInFile(unknownAddr)
-	test.That(t, err, test.ShouldBeError, errNoDevice)
-	test.That(t, device, test.ShouldBeNil)
-
-	// Test File read error
-	g.dataFile.Close()
-	_, err = g.searchForDeviceInFile(testDeviceAddr)
-	test.That(t, err, test.ShouldNotBeNil)
-	test.That(t, err.Error(), test.ShouldContainSubstring, "failed to read device info from file")
-}
-
-func TestUpdateDeviceInfo(t *testing.T) {
-	g := setupTestGateway(t)
-
-	newAppSKey := []byte{
-		0x55, 0x72, 0x40, 0x4C,
-		0x69, 0x6E, 0x6B, 0x4C,
-		0x6F, 0x52, 0x61, 0x32,
-		0x31, 0x30, 0x32, 0x23,
-	}
-
-	newDevAddr := []byte{0xe2, 0x73, 0x65, 0x67}
-
-	// Test 1: Successful update
-	validInfo := &deviceInfo{
-		DevEUI:  fmt.Sprintf("%X", testDevEUI), // matches dev EUI on the gateway map
-		DevAddr: fmt.Sprintf("%X", newDevAddr),
-		AppSKey: fmt.Sprintf("%X", newAppSKey),
-	}
-
-	device, err := g.updateDeviceInfo(validInfo)
-	test.That(t, err, test.ShouldBeNil)
-	test.That(t, device, test.ShouldNotBeNil)
-	test.That(t, device.NodeName, test.ShouldEqual, testNodeName)
-	test.That(t, device.AppSKey, test.ShouldResemble, newAppSKey)
-	test.That(t, device.Addr, test.ShouldResemble, newDevAddr)
-
-	// If the device is not found in the map should return error
-	g.devices = make(map[string]*node.Node) // clear devices map
-	_, err = g.updateDeviceInfo(validInfo)
-	test.That(t, err, test.ShouldNotBeNil)
-	test.That(t, err.Error(), test.ShouldContainSubstring, "could not find the matching device in device map")
 }
 
 func TestConvertTo32Bit(t *testing.T) {
